fix: resolve flashing of embedded youtube videos
